backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pyjwt==2.8.0
passlib==1.7.4
sqlalchemy==2.0.23
python-jose==3.3.0
requests==2.31.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
google-api-python-client==2.108.0

# MySQL Support
pymysql==1.1.0
cryptography==3.4.8

# Database Migrations
alembic==1.12.1

# Lemon Squeezy Integration
cryptography==3.4.8
